+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/job/JobInitializer.java
@@ -39,7 +39,7 @@ public class JobInitializer {
                 LOGGER.error("No runnable job type found for " + job.getJobType().getName());
             }
             if (runnableJob != null) {
-                LOGGER.info("Starting job with ID " + job.getId() + " for user " + job.getUser().getSubjectName());
+                LOGGER.info("Starting job with ID " + job.getId() + " for user " + job.getUser().getUsername());
                 taskExecutor.execute(runnableJob);
             }
         }