+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/scheduler/job/summarystatistics/SummaryStatisticsCreatorJob.java
@@ -5,9 +5,12 @@ import java.time.ZoneOffset;
 import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.Date;
+import java.util.Hashtable;
 import java.util.List;
+import java.util.Map;
 import java.util.TimeZone;
 import java.util.concurrent.ExecutionException;
+import java.util.stream.Collectors;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -26,8 +29,11 @@ import org.springframework.web.context.support.SpringBeanAutowiringSupport;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
+import gov.healthit.chpl.dao.CertificationStatusEventDAO;
+import gov.healthit.chpl.dao.CertifiedProductDAO;
 import gov.healthit.chpl.dao.statistics.SummaryStatisticsDAO;
-import gov.healthit.chpl.domain.DateRange;
+import gov.healthit.chpl.dto.CertificationStatusEventDTO;
+import gov.healthit.chpl.dto.CertifiedProductDetailsDTO;
 import gov.healthit.chpl.entity.SummaryStatisticsEntity;
 import gov.healthit.chpl.exception.EntityCreationException;
 import gov.healthit.chpl.exception.EntityRetrievalException;
@@ -51,6 +57,12 @@ public class SummaryStatisticsCreatorJob extends QuartzJob {
     private SummaryStatisticsDAO summaryStatisticsDAO;
 
     @Autowired
+    private CertifiedProductDAO certifiedProductDAO;
+
+    @Autowired
+    private CertificationStatusEventDAO certificationStatusEventDAO;
+
+    @Autowired
     private JpaTransactionManager txManager;
 
     @Autowired
@@ -61,25 +73,19 @@ public class SummaryStatisticsCreatorJob extends QuartzJob {
     }
 
     @Override
-    public void execute(final JobExecutionContext jobContext) throws JobExecutionException {
+    public void execute(JobExecutionContext jobContext) throws JobExecutionException {
         LOGGER.info("********* Starting the Summary Statistics Creation job. *********");
         try {
             SpringBeanAutowiringSupport.processInjectionBasedOnCurrentContext(this);
 
-            Boolean generateCsv = Boolean.valueOf(jobContext.getMergedJobDataMap().getString("generateCsvFile"));
-            Date startDate = getStartDate();
-            if (startDate == null) {
-                throw new RuntimeException("Could not obtain the startDate.");
-            }
-            Date endDate = new Date();
-            Integer numDaysInPeriod = Integer.valueOf(env.getProperty("summaryEmailPeriodInDays").toString());
+            LOGGER.info("Getting all listings.");
+            List<CertifiedProductDetailsDTO> allListings = certifiedProductDAO.findAll();
+            LOGGER.info("Completing getting all listings.");
 
-            EmailStatistics emailBodyStats = emailStatisticsCreator.getStatistics();
+            EmailStatistics emailBodyStats = emailStatisticsCreator.getStatistics(allListings);
+            saveSummaryStatistics(emailBodyStats);
 
-            if (generateCsv) {
-                createSummaryStatisticsFile(startDate, endDate, numDaysInPeriod);
-            }
-            saveSummaryStatistics(emailBodyStats, endDate);
+            createSummaryStatisticsFile(allListings, jobContext);
 
         } catch (Exception e) {
             LOGGER.error("Caught unexpected exception: " + e.getMessage(), e);
@@ -87,47 +93,51 @@ public class SummaryStatisticsCreatorJob extends QuartzJob {
         LOGGER.info("********* Completed the Summary Statistics Creation job. *********");
     }
 
-    private void createSummaryStatisticsFile(final Date startDate, final Date endDate, final Integer numDaysInPeriod)
+
+    @SuppressWarnings("checkstyle:linelength")
+    private void createSummaryStatisticsFile(List<CertifiedProductDetailsDTO> allListings, JobExecutionContext jobContext)
             throws InterruptedException, ExecutionException {
+
+
+        if (!isGenerateStatisticsFlagOn(jobContext)) {
+            return;
+        }
+
+        Date startDate = getStartDate();
+        if (startDate == null) {
+            throw new RuntimeException("Could not obtain the startDate.");
+        }
+        Date endDate = new Date();
+        Integer numDaysInPeriod = Integer.valueOf(env.getProperty("summaryEmailPeriodInDays").toString());
+
         List<CsvStatistics> csvStats = new ArrayList<CsvStatistics>();
-        Calendar startDateCal = Calendar.getInstance(TimeZone.getTimeZone(ZoneOffset.UTC));
-        startDateCal.setTime(startDate);
         Calendar endDateCal = Calendar.getInstance(TimeZone.getTimeZone(ZoneOffset.UTC));
         endDateCal.setTime(startDate);
-        endDateCal.add(Calendar.DATE, numDaysInPeriod);
+
+        Map<Long, List<CertificationStatusEventDTO>> statusesForAllListings = getAllStatusesForAllListings();
 
         while (endDate.compareTo(endDateCal.getTime()) >= 0) {
-            LOGGER.info("Getting csvRecord for start date " + startDateCal.getTime().toString() + " end date "
-                    + endDateCal.getTime().toString());
-            DateRange csvRange = new DateRange(startDateCal.getTime(), new Date(endDateCal.getTimeInMillis()));
             CsvStatistics historyStat = new CsvStatistics();
-            historyStat.setDateRange(csvRange);
-            historyStat = historicalStatisticsCreator.getStatistics(csvRange);
+            historyStat.setEndDate(endDateCal.getTime());
+            historyStat = historicalStatisticsCreator.getStatistics(allListings, statusesForAllListings, endDateCal.getTime());
             csvStats.add(historyStat);
-            LOGGER.info("Finished getting csvRecord for start date " + startDateCal.getTime().toString() + " end date "
-                    + endDateCal.getTime().toString());
-            startDateCal.add(Calendar.DATE, numDaysInPeriod);
-            endDateCal.setTime(startDateCal.getTime());
             endDateCal.add(Calendar.DATE, numDaysInPeriod);
         }
-        LOGGER.info("Finished getting statistics");
 
-        LOGGER.info("Writing statistics CSV");
         StatsCsvFileWriter csvFileWriter = new StatsCsvFileWriter();
         csvFileWriter.writeCsvFile(env.getProperty("downloadFolderPath") + File.separator
                 + env.getProperty("summaryEmailName", "summaryStatistics.csv"), csvStats);
 
         new File(env.getProperty("downloadFolderPath") + File.separator
                 + env.getProperty("summaryEmailName", "summaryStatistics.csv"));
-        LOGGER.info("Completed statistics CSV");
     }
 
-    private String getJson(final EmailStatistics statistics) throws JsonProcessingException {
+    private String getJson(EmailStatistics statistics) throws JsonProcessingException {
         ObjectMapper mapper = new ObjectMapper();
         return mapper.writeValueAsString(statistics);
     }
 
-    public void saveSummaryStatistics(final EmailStatistics statistics, final Date endDate)
+    public void saveSummaryStatistics(EmailStatistics statistics)
             throws JsonProcessingException, EntityCreationException, EntityRetrievalException {
 
         // We need to manually create a transaction in this case because of how AOP works. When a method is
@@ -145,7 +155,7 @@ public class SummaryStatisticsCreatorJob extends QuartzJob {
                     summaryStatisticsDAO.deleteAll();
 
                     SummaryStatisticsEntity entity = new SummaryStatisticsEntity();
-                    entity.setEndDate(endDate);
+                    entity.setEndDate(new Date());
                     entity.setSummaryStatistics(getJson(statistics));
                     summaryStatisticsDAO.create(entity);
                 } catch (Exception e) {
@@ -169,12 +179,25 @@ public class SummaryStatisticsCreatorJob extends QuartzJob {
             return startDateCalendar.getTime();
         }
         for (int i = 0; i <= 6; i++) {
-            startDateCalendar.add(Calendar.DATE, (-1));
+            startDateCalendar.add(Calendar.DATE, 1);
             if (startDateCalendar.get(Calendar.DAY_OF_WEEK) == dow) {
                 return startDateCalendar.getTime();
             }
         }
         return null;
+    }
+
+    private Map<Long, List<CertificationStatusEventDTO>> getAllStatusesForAllListings() {
+        Map<Long, List<CertificationStatusEventDTO>> map = certificationStatusEventDAO.findAll().stream()
+                .collect(Collectors.groupingBy(CertificationStatusEventDTO::getCertifiedProductId));
+
+        Map<Long, List<CertificationStatusEventDTO>> syncdMap = new Hashtable<Long, List<CertificationStatusEventDTO>>();
+        syncdMap.putAll(map);
+        return syncdMap;
+    }
+
+    private Boolean isGenerateStatisticsFlagOn(JobExecutionContext jobContext) {
+        return Boolean.valueOf(jobContext.getMergedJobDataMap().getString("generateCsvFile"));
     }
 
 }